Set bank switching logic

+++ Hardware/IMapper.cs
@@ -16,7 +16,7 @@ public interface IMapper
     public bool IsCpuWrite(ushort address);
     public int? CpuRead(ushort address);
     public int? CpuWrite(ushort address, byte data);
-    public bool PpuRead(ushort address, out ushort mappedAddress);
+    public bool PpuRead(ushort address, out int mappedAddress);
     public bool PpuWrite(ushort address, out ushort mappedAddress);
 }
 
@@ -51,7 +51,7 @@ public abstract class AbstractMapper : IMapper
 
     public abstract int? CpuWrite(ushort address, byte data);
 
-    public abstract bool PpuRead(ushort address, out ushort mappedAddress);
+    public abstract bool PpuRead(ushort address, out int mappedAddress);
 
     public abstract bool PpuWrite(ushort address, out ushort mappedAddress);
 }
