+++ Hardware/Mappers/Nrom.cs
@@ -33,7 +33,7 @@ public class NRom : AbstractMapper
         return null;
     }
 
-    public override bool PpuRead(ushort address, out ushort mappedAddress)
+    public override bool PpuRead(ushort address, out int mappedAddress)
     {
         if (address <= 0x1FFF)
         {
